Moved placement rules to a package of their own, added two game flow rule (stubs), added another event type to UnoGameFlowRule.
