+++ src/main/java/com/github/markozajc/juno/rules/types/UnoGameFlowRule.java
@@ -1,11 +1,14 @@
 package com.github.markozajc.juno.rules.types;
 
+import com.github.markozajc.juno.cards.UnoCard;
 import com.github.markozajc.juno.game.UnoGame;
 import com.github.markozajc.juno.hands.UnoHand;
 import com.github.markozajc.juno.rules.UnoRule;
 
 public interface UnoGameFlowRule extends UnoRule {
 
-	public void onTurn(UnoHand hand, UnoGame game);
+	public default void turnInitialization(UnoHand hand, UnoGame game) {}
+
+	public default void afterHandDecision(UnoHand hand, UnoGame game, UnoCard decidedCard) {}
 
 }
